split out public class SmallSystem from Polynomials.ts to keep poly solver internals out of the API (and avoid circular dependency)
