+++ core/geometry/src/curve/Arc3d.ts
@@ -23,7 +23,8 @@ import { Transform } from "../geometry3d/Transform";
 import { XYAndZ } from "../geometry3d/XYZProps";
 import { Matrix4d } from "../geometry4d/Matrix4d";
 import { Point4d } from "../geometry4d/Point4d";
-import { SineCosinePolynomial, SmallSystem, TrigPolynomial } from "../numerics/Polynomials";
+import { SineCosinePolynomial, TrigPolynomial } from "../numerics/Polynomials";
+import { SmallSystem } from "../numerics/SmallSystem";
 import { CurveChain } from "./CurveCollection";
 import { CurveExtendMode, CurveExtendOptions, VariantCurveExtendParameter } from "./CurveExtendMode";
 import { CurveIntervalRole, CurveLocationDetail, CurveSearchStatus } from "./CurveLocationDetail";
