+++ core/geometry/src/curve/internalContexts/CurveCurveIntersectXY.ts
@@ -26,7 +26,8 @@ import { Matrix4d } from "../../geometry4d/Matrix4d";
 import { Point4d } from "../../geometry4d/Point4d";
 import { UnivariateBezier } from "../../numerics/BezierPolynomials";
 import { CurveCurveIntersectionXYRRToRRD, Newton2dUnboundedWithDerivative } from "../../numerics/Newton";
-import { AnalyticRoots, SmallSystem, TrigPolynomial } from "../../numerics/Polynomials";
+import { AnalyticRoots, TrigPolynomial } from "../../numerics/Polynomials";
+import { SmallSystem } from "../../numerics/SmallSystem";
 import { Arc3d } from "../Arc3d";
 import { CurveChainWithDistanceIndex } from "../CurveChainWithDistanceIndex";
 import { CurveCollection } from "../CurveCollection";
