+++ core/geometry/src/curve/internalContexts/CurveCurveIntersectXYZ.ts
@@ -15,7 +15,8 @@ import { RecurseToCurvesGeometryHandler } from "../../geometry3d/GeometryHandler
 import { Plane3dByOriginAndUnitNormal } from "../../geometry3d/Plane3dByOriginAndUnitNormal";
 import { Vector2d } from "../../geometry3d/Point2dVector2d";
 import { Point3d, Vector3d } from "../../geometry3d/Point3dVector3d";
-import { SmallSystem, TrigPolynomial } from "../../numerics/Polynomials";
+import { TrigPolynomial } from "../../numerics/Polynomials";
+import { SmallSystem } from "../../numerics/SmallSystem";
 import { Arc3d } from "../Arc3d";
 import { CurveChainWithDistanceIndex } from "../CurveChainWithDistanceIndex";
 import { CurveCollection } from "../CurveCollection";
