+++ core/geometry/src/test/curve/BuildingCodeOffsetOps.ts
@@ -4,16 +4,18 @@
 *--------------------------------------------------------------------------------------------*/
 
 import { Arc3d } from "../../curve/Arc3d";
+import { CurvePrimitive } from "../../curve/CurvePrimitive";
+import { LineSegment3d } from "../../curve/LineSegment3d";
+import { Loop } from "../../curve/Loop";
+import { Path } from "../../curve/Path";
 import { Geometry } from "../../Geometry";
 import { AngleSweep } from "../../geometry3d/AngleSweep";
-import { Point3d, Vector3d } from "../../geometry3d/Point3dVector3d";
-import { LineSegment3d } from "../../curve/LineSegment3d";
-import { AnalyticRoots, SmallSystem } from "../../numerics/Polynomials";
-import { Vector2d } from "../../geometry3d/Point2dVector2d";
-import { Path } from "../../curve/Path";
-import { Loop } from "../../curve/Loop";
-import { CurvePrimitive } from "../../curve/CurvePrimitive";
 import { GrowableFloat64Array } from "../../geometry3d/GrowableFloat64Array";
+import { Vector2d } from "../../geometry3d/Point2dVector2d";
+import { Point3d, Vector3d } from "../../geometry3d/Point3dVector3d";
+import { AnalyticRoots } from "../../numerics/Polynomials";
+import { SmallSystem } from "../../numerics/SmallSystem";
+
 /**
  * Assorted static methods for constructing fragmentary and complete curves for offsets from linestrings.
  * * Primary method is the static edgeByEdgeOffsetFromPoints
