+++ core/geometry/src/topology/InsertAndRetriangulateContext.ts
@@ -11,7 +11,7 @@ import { Geometry, PolygonLocation } from "../Geometry";
 import { Point3d } from "../geometry3d/Point3dVector3d";
 import { PolygonOps } from "../geometry3d/PolygonOps";
 import { Ray3d } from "../geometry3d/Ray3d";
-import { SmallSystem } from "../numerics/Polynomials";
+import { SmallSystem } from "../numerics/SmallSystem";
 import { HalfEdge, HalfEdgeGraph, HalfEdgeMask } from "./Graph";
 import { MarkedEdgeSet } from "./HalfEdgeMarkSet";
 import { PointSearchContext, RayClassification } from "./HalfEdgePointInGraphSearch";
