+++ core/geometry/src/topology/Merging.ts
@@ -12,15 +12,15 @@ import { CurveLocationDetail } from "../curve/CurveLocationDetail";
 import { LineSegment3d } from "../curve/LineSegment3d";
 import { Geometry } from "../Geometry";
 import { Angle } from "../geometry3d/Angle";
+import { MultiLineStringDataVariant } from "../geometry3d/IndexedXYZCollection";
 import { Point2d, Vector2d } from "../geometry3d/Point2dVector2d";
 import { Range3d } from "../geometry3d/Range";
 import { ClusterableArray } from "../numerics/ClusterableArray";
-import { SmallSystem } from "../numerics/Polynomials";
+import { SmallSystem } from "../numerics/SmallSystem";
 import { HalfEdge, HalfEdgeGraph, HalfEdgeMask } from "./Graph";
 import { HalfEdgePriorityQueueWithPartnerArray } from "./HalfEdgePriorityQueue";
 import { RegularizationContext } from "./RegularizeFace";
 import { Triangulator } from "./Triangulation";
-import { MultiLineStringDataVariant } from "../geometry3d/IndexedXYZCollection";
 
 export class GraphSplitData {
   public numUpEdge = 0;
